moved WorstLinesChart logic to separate component
